refactor: Update task component to display formatted due date
This commit modifies the task component to display the due date in a formatted way using the Angular `DatePipe`. The `task.component.html` file is updated to apply the `date` pipe to the `task.dueDate` property, formatting it as the 'fullDate'. This change improves the readability of the due date for users.